requirements.txt
pypdf>=4.0.0
ftfy>=6.1.1
openai>=1.30.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pgvector>=0.4.1
reportlab>=4.0.0
pytest>=8.0.0
